refactor: 영화 목록을 MovieCardList 컴포넌트로 분리
